+++ abci/did/dataStruct.go
@@ -127,6 +127,7 @@ type Response struct {
 	IdpID            string  `json:"idp_id"`
 	ValidProof       *bool   `json:"valid_proof"`
 	ValidIal         *bool   `json:"valid_ial"`
+	ValidSignature   *bool   `json:"valid_signature"`
 }
 
 type CreateIdpResponseParam struct {
@@ -465,9 +466,10 @@ type TimeOutRequestParam struct {
 }
 
 type ResponseValid struct {
-	IdpID      string `json:"idp_id"`
-	ValidProof *bool  `json:"valid_proof"`
-	ValidIal   *bool  `json:"valid_ial"`
+	IdpID          string `json:"idp_id"`
+	ValidProof     *bool  `json:"valid_proof"`
+	ValidIal       *bool  `json:"valid_ial"`
+	ValidSignature *bool  `json:"valid_signature"`
 }
 
 type GetDataSignatureParam struct {